move towards using a helper utilities file
